Plotly embedding figures

+++ bacpipe/core/workflows.py
@@ -19,9 +19,11 @@ from bacpipe.embedding_evaluation.visualization.dashboard import (
 )
 
 from bacpipe.embedding_evaluation.visualization.visualize import (
+    visualise_results_across_models,
+)
+from bacpipe.embedding_evaluation.visualization.visualize_embeddings import (
     plot_comparison,
     plot_embeddings,
-    visualise_results_across_models,
     EmbedAndLabelLoader,
 )
 from bacpipe.embedding_evaluation.label_embeddings import (
@@ -68,7 +70,7 @@ def play(config=config, settings=settings, bool_save_logs=False):
                                                    model_names=config.models)
     overwrite, dashboard = config.overwrite, config.dashboard
 
-    if config.audio_dir == 'bacpipe/tests/test_data':
+    if config.audio_dir == 'bacpipe/tests/test_data' or settings.testing:
         with pkg_resources.path(
             __package__.split('.')[0] + ".tests.test_data", ""
             ) as audio_dir:
@@ -344,17 +346,38 @@ def model_specific_embedding_creation(audio_dir, dim_reduction_model, models, **
         dictionary containing the loader objects for each model
     """
     loader_dict = {}
+    remove_models_from_list = []
     for model_name in models:
-        loader_dict[model_name] = run_pipeline_for_model(
-            model_name=model_name,
-            dim_reduction_model=dim_reduction_model,
-            audio_dir=audio_dir,
-            **kwargs,
-        )
+        try:
+            loader_dict[model_name] = run_pipeline_for_model(
+                model_name=model_name,
+                dim_reduction_model=dim_reduction_model,
+                audio_dir=audio_dir,
+                **kwargs,
+            )
+        except AssertionError as e:
+            remove_models_from_list.append(model_name)
+            if kwargs['already_computed']:
+                logger.exception(
+                    f"Bacpipe was not able to process {model_name} because {e}. "
+                    f"Because `already_computed` is True, it looks like {model_name} "
+                    "didn't fully finish on the last run. "
+                    "Bacpipe will continue without this model so that the rest of "
+                    "the processing can still be completed. "
+                    "To ensure this model get's processed, set `already_computed` to False."
+                    )
+            else:
+                logger.exception(
+                    f"Bacpipe was not able to process {model_name} because {e}."
+                )
+    if len(remove_models_from_list) > 0:
+        for model in remove_models_from_list:
+            models.remove(model)
     return loader_dict
 
 def model_specific_evaluation(
-    loader_dict, evaluation_task, class_configs, models, **kwargs
+    loader_dict, evaluation_task, class_configs, 
+    models, dim_reduction_model=False, **kwargs
 ):
     """
     Perform evaluation of the embeddings using the specified
@@ -382,19 +405,30 @@ def model_specific_evaluation(
     """
     for model_name in models:
         paths = get_paths(model_name)
-        loader_dict[model_name].check_if_default_clfier_should_be_run(
-            paths, **kwargs
-            )
+        if loader_dict[model_name].classifier_should_be_run(paths, **kwargs):
+            embed = Embedder(model_name, loader_dict[model_name], **kwargs)
+            if hasattr(embed.model, 'classifier_predictions'):
+                embed.classifier.run_default_classifier(
+                    loader_dict[model_name]
+                    )
                 
         if not evaluation_task in ["None", [], False]:
             embeds = loader_dict[model_name].embeddings()
             try:
-                ground_truth = ground_truth_by_model(paths, model_name, **kwargs)
+                ground_truth = ground_truth_by_model(model_name, paths=paths, **kwargs)
             except FileNotFoundError as e:
                 ground_truth = None
 
+
+
+        ####################################################################
+        ############      PROBING OF EMBEDDINGS THROUGH       ##############
+        ############      LINEAR AND KNN CLASSIFICATION       ##############
+        ############            SEE SETTINGS.YAML             ##############
+        ####################################################################
+        
         if "classification" in evaluation_task and not ground_truth is None:
-            print(
+            logger.info(
                 "\nTraining classifier to evaluate " f"{model_name.upper()} embeddings"
             )
 
@@ -421,8 +455,13 @@ def model_specific_evaluation(
                         paths, class_embeds, **class_config, **kwargs
                     )
 
+        ####################################################################
+        ############      CLUSTERING OF EMBEDDINGS THROUGH    ##############
+        ######      KMEANS (AND WHATEVER SPECIFIED IN SETTINGS.YAML)   #####
+        ####################################################################
+                    
         if "clustering" in evaluation_task:
-            print(
+            logger.info(
                 "\nGenerating clusterings to evaluate "
                 f"{model_name.upper()} embeddings"
             )
@@ -493,9 +532,10 @@ def run_pipeline_for_model(
 
     if not dim_reduction_model in ["None", False]:
 
-        assert len(loader_embeddings.files) > 1, (
+        assert len(loader_embeddings.files) > 1, logger.exception(
             "Too few files to perform dimensionality reduction. "
-            + "Are you sure you have selected the right data?"
+            "Are you sure you have selected the right data? "
+            f"Will continue without {model_name}."
         )
         loader_dim_reduced = generate_embeddings(
             model_name=model_name,
@@ -515,7 +555,7 @@ def run_pipeline_for_model(
                 " Skipping visualization generation."
             )
         else:
-            print(
+            logger.info(
                 "### Generating visualizations of embeddings using "
                 f"{dim_reduction_model}. Plots are saved in "
                 f"{loader_dim_reduced.embed_dir} ###"
@@ -523,25 +563,32 @@ def run_pipeline_for_model(
             vis_loader = EmbedAndLabelLoader(
                 dim_reduction_model=dim_reduction_model, **kwargs
             )
-            plot_embeddings(
-                vis_loader,
-                paths=paths,
-                model_name=loader_dim_reduced.model_name,
-                dim_reduction_model=dim_reduction_model,
-                bool_plot_centroids=False,
-                label_by="time_of_day",
-                **kwargs,
-            )
+            try:
+                plot_embeddings(
+                    vis_loader,
+                    paths=paths,
+                    model_name=loader_dim_reduced.model_name,
+                    dim_reduction_model=dim_reduction_model,
+                    bool_plot_centroids=False,
+                    label_by="time_of_day",
+                    **kwargs,
+                )
+            except AssertionError as e:
+                logger.exception(
+                    "Plotting of embeddings has failed. Continuing with processing "
+                    f"embeddings, but this will cause evaluation problems later on. {e}"
+                )
+                
     return loader_embeddings
 
 
 def generate_embeddings(avoid_pipelined_gpu_inference=False, **kwargs):
     if "dim_reduction_model" in kwargs:
-        print(
+        logger.info(
             f"\n\n\n###### Generating embeddings using {kwargs['dim_reduction_model'].upper()} ######\n"
         )
     elif "model_name" in kwargs:
-        print(
+        logger.info(
             f"\n\n\n###### Generating embeddings using {kwargs['model_name'].upper()} ######\n"
         )
     else:
@@ -583,10 +630,14 @@ def generate_embeddings(avoid_pipelined_gpu_inference=False, **kwargs):
         return ld
     except KeyboardInterrupt:
         if ld.embed_dir.exists() and ld.rm_embedding_on_keyboard_interrupt:
-            print("KeyboardInterrupt: Exiting and deleting created embeddings.")
-            import shutil
+            all_files = list(Path(ld.embed_dir).rglob('*'))
+            if len(all_files) < 25:
+                logger.info(f"KeyboardInterrupt: Exiting and deleting created {ld.embed_dir}.")
+                import shutil
 
-            shutil.rmtree(ld.embed_dir)
+                shutil.rmtree(ld.embed_dir)
+            else:
+                logger.info(f"KeyboardInterrupt: Exiting and but not deleting {ld.embed_dir}.")
         import sys
 
         sys.exit()
